requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
deepface==0.0.79
faiss-cpu==1.13.0
fastapi==0.122.0
h11==0.16.0
idna==3.11
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
